test: Split test files
Move type specific unit tests from uint_test.cpp to uint256_test.cpp.
The MSVC 2017 has issues with compiling big uint_test.cpp.
